- feat: Converted the project into CMake for Cross-Platform support - Tested on WSL Ubuntu 24.04.amd64 - Fixed the project tree and subfolders - included modern GLFW library for both linux - windows in Extern - fix: some issues related to cross-platform (windows-linux)

+++ LibImageUI/ImGuiFileDialog.cpp
@@ -26,7 +26,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include "stdafx.h"
+#include "Stdafx.h"
 
 #include "ImGuiFileDialog.h"
 
@@ -3995,7 +3995,7 @@ void IGFD::FileDialog::m_DisplayPathPopup(ImVec2 vSize) {
 
                         if (ImGui::TableNextColumn())  // file name
                         {
-                            if (ImGui::Selectable(infos_ptr->fileNameExt.c_str(), &selected, ImGuiSelectableFlags_SpanAllColumns | ImGuiSelectableFlags_SpanAvailWidth)) {
+                            if (ImGui::Selectable(infos_ptr->fileNameExt.c_str(), &selected, (int)ImGuiSelectableFlags_SpanAllColumns | (int)ImGuiSelectableFlags_SpanAvailWidth)) {
                                 fdi.SetCurrentPath(fdi.ComposeNewPath(fdi.GetCurrentPopupComposedPath()));
                                 fdi.pathClicked = fdi.SelectDirectory(infos_ptr);
                                 ImGui::CloseCurrentPopup();
